requirements/production.txt
# PRECAUTION: avoid production dependencies that aren't in development

-r base.txt

gunicorn==23.0.0  # https://github.com/benoitc/gunicorn
psycopg[c]==3.2.6  # https://github.com/psycopg/psycopg
sentry-sdk==2.22.0  # https://github.com/getsentry/sentry-python

# Django
# ------------------------------------------------------------------------------
django-anymail==12.0  # https://github.com/anymail/django-anymail

# Storages
django-storages==1.14.4
boto3==1.36.20
